Fix issues in history screen

+++ app/src/main/java/com/amitnadiger/myinvestment/viewModel/FinHistoryViewModel.kt
@@ -68,7 +68,7 @@ class FinHistoryViewModel(application: Application): ViewModel() {
     }
 
     fun deleteFinProduct(accountNum: String) {
-        repository.deleteProduct(accountNum.toLong())
+        repository.deleteProduct(accountNum)
     }
 
     fun deleteAllFromHistory() {
